modularized more code
