+++ src/javascripts/components/form/eat.js
@@ -28,9 +28,4 @@ const buttonFunc = (e) => {
   }
 };
 
-const buttonHandleEvent = () => {
-  document.querySelector('#healthyButton').addEventListener('click', buttonFunc);
-  document.querySelector('#badButton').addEventListener('click', buttonFunc);
-};
-
-export default buttonHandleEvent;
+export default buttonFunc;
